fix some warnings
git-svn-id: https://svn.code.sf.net/p/pinmame/code/trunk@4132 6f51dfdd-75df-4273-8f3b-818b1e6aa8bf

+++ src/wpc/playsnd.c
@@ -356,7 +356,7 @@ static WRITE_HANDLER(ay8910_1_porta_w)	{
 }
 struct AY8910interface play4s_8910Int = {
 	2,			/* 2 chips */
-	3579545.0/2.,	/* 1.79 MHz */
+	(int)(3579545.0/2.),	/* 1.79 MHz */
 	{ MIXER(50,MIXER_PAN_LEFT), MIXER(50,MIXER_PAN_RIGHT) },	/* Volume */
 	{ 0, 0 },
 	{ 0, 0 },
